tidy up and renaming
