+++ test/functional/integration/context/context.it-spec.ts
@@ -32,7 +32,7 @@ const contextVision = 'contextVision';
 const contextImpact = 'contextImpact';
 const contextWho = 'contextWho';
 const refName = 'refName';
-const refUri = 'https://test.cherrytwist.org/';
+const refUri = 'https://test.alekm.io/';
 const tagsArray = ['tag1', 'tag2'];
 let challengeContextData = '';
 let challengeRefName = '';
